Fixed type errors
